refactor: move all things related to claim to a custom hook
